intergrate admin side

+++ backend/Models/OrderDetail.cs
@@ -2,7 +2,7 @@ namespace backend.Models;
 
 public class OrderDetail : BaseModel
 {
-  public int SalePrice { get; set; }
+  public double SalePrice { get; set; }
   public int Quantity { get; set; }
   public Guid OrderId { get; set; }
   public Guid ProductId { get; set; }
